+++ fe/src/components/Common/Dropdown/index.tsx
@@ -2,16 +2,16 @@ import { MutableRefObject, ReactNode } from 'react';
 
 import * as S from './style';
 
-interface PopUpProps {
+interface DropdownProps {
   parentComponent: MutableRefObject<HTMLDivElement>;
   isOpen: boolean;
   onClose: (event: MouseEvent) => void;
   children: ReactNode;
 }
 
-export default function PopUp({ parentComponent, children, isOpen, onClose }: PopUpProps) {
+export default function Dropdown({ parentComponent, children, isOpen, onClose }: DropdownProps) {
   if (!isOpen) return null;
-  return <S.Container>{children}</S.Container>;
+  return <S.Dropdown>{children}</S.Dropdown>;
 }
 
 // 여러 parent등록 가능하게
